code/requirements.txt
asgiref==3.5.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.1.14
django-appconf==1.0.5
django-baton==2.3.0
django-compressor==3.1
django-fsm==2.8.0
django-libsass==0.9
djangorestframework==3.13.1
graphviz==0.20
httplib2==0.20.4
idna==3.3
libsass==0.21.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
pytz==2022.1
rcssmin==1.1.0
requests==2.27.1
rjsmin==1.2.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
